sql_executors/requirements.txt
tabulate==0.8.5
redis==3.3.11
asyncio==3.4.3
sentry-sdk==0.13.4
mysql-connector==2.2.9